+++ packages/yourloops/components/common.tsx
@@ -27,9 +27,9 @@
 
 import { makeStyles, Theme } from "@material-ui/core/styles";
 
-export const commonTeamStyles = makeStyles((theme: Theme) => ({
+export const commonComponentStyles = makeStyles((theme: Theme) => ({
   button: {
-    textTransform: "none",
+    textTransform: "capitalize",
   },
   categoryHeader: {
     "display": "flex",