+++ packages/yourloops/components/patient/row.tsx
@@ -28,6 +28,7 @@
 import React from "react";
 import _ from "lodash";
 import { useTranslation } from "react-i18next";
+import moment from "moment-timezone";
 
 import { makeStyles, Theme } from "@material-ui/core/styles";
 import Tooltip from "@material-ui/core/Tooltip";
@@ -94,30 +95,42 @@ function PatientRow(props: PatientElementProps): JSX.Element {
   const authHook = useAuth();
   const teamHook = useTeam();
   const isUserHcp = authHook.user?.isUserHcp();
-  const patientIsMonitored = patient.remoteMonitoring !== null;
+  const patientIsMonitored = patient.monitoring !== null;
   const classes = patientListStyle();
   const patientListCommonClasses = patientListCommonStyle();
-  const [medicalData, setMedicalData] = React.useState<MedicalData | null | undefined>(patient.medicalData);
+  const [medicalData, setMedicalData] = React.useState<MedicalData | null | undefined>(patient.metadata.medicalData);
   const [tooltipText, setTooltipText] = React.useState<string>("");
   const rowRef = React.createRef<HTMLTableRowElement>();
 
   const userId = patient.userid;
-  const email = patient.username;
+  const email = patient.profile.email;
   const isFlagged = flagged.includes(userId);
-  const patientFullName = patient.fullName;
+  const patientFullName = patient.profile.fullName;
 
   const computeRowInformation = () => {
     const mediumCellWithAlertClasses = `${classes.typography} ${patientListCommonClasses.mediumCell} ${classes.alert}`;
     const mediumCellWithClasses = `${classes.typography} ${patientListCommonClasses.mediumCell}`;
-    const timeSpentAwayFromTargetActive = patient.alarm?.timeSpentAwayFromTargetActive ?? false;
-    const frequencyOfSevereHypoglycemiaActive = patient.alarm?.frequencyOfSevereHypoglycemiaActive ?? false;
-    const nonDataTransmissionActive = patient.alarm?.nonDataTransmissionActive ?? false;
+    const timeSpentAwayFromTargetActive = patient.metadata.alarm?.timeSpentAwayFromTargetActive ?? false;
+    const frequencyOfSevereHypoglycemiaActive = patient.metadata.alarm?.frequencyOfSevereHypoglycemiaActive ?? false;
+    const nonDataTransmissionActive = patient.metadata.alarm?.nonDataTransmissionActive ?? false;
+    let patientRemoteMonitoring;
+    if (patient.monitoring?.enabled) {
+      if (patient.monitoring.monitoringEnd) {
+        const enDate = moment.utc(patient.monitoring.monitoringEnd).format(moment.localeData().longDateFormat("ll")).toString();
+        patientRemoteMonitoring = `${t("yes")}\n(${t("until")} ${enDate})`;
+      } else {
+        patientRemoteMonitoring = t("yes");
+      }
+    } else {
+      patientRemoteMonitoring = t("no");
+    }
+
     let patientFullNameClasses = `${classes.typography} ${patientListCommonClasses.largeCell}`;
     let timeSpentAwayFromTargetRateClasses = mediumCellWithClasses;
     let frequencyOfSevereHypoglycemiaRateClasses = mediumCellWithClasses;
     let dataNotTransferredRateClasses = mediumCellWithClasses;
     if (isUserHcp) {
-      const dataNotTransferredActive = patient.alarm?.nonDataTransmissionRate ?? false;
+      const dataNotTransferredActive = patient.metadata.alarm?.nonDataTransmissionRate ?? false;
       const hasAlert = timeSpentAwayFromTargetActive || frequencyOfSevereHypoglycemiaActive || nonDataTransmissionActive;
       patientFullNameClasses = hasAlert ? `${classes.typography} ${classes.alert} ${patientListCommonClasses.largeCell}` : `${classes.typography} ${patientListCommonClasses.largeCell}`;
       timeSpentAwayFromTargetRateClasses = timeSpentAwayFromTargetActive ? mediumCellWithAlertClasses : mediumCellWithClasses;
@@ -125,8 +138,8 @@ function PatientRow(props: PatientElementProps): JSX.Element {
       dataNotTransferredRateClasses = dataNotTransferredActive ? mediumCellWithAlertClasses : mediumCellWithClasses;
     }
     return {
-      patientSystem: patient.system ?? trNA,
-      patientRemoteMonitoring: patient.remoteMonitoring ? `${t("yes")}\n(${t("since")} ${patient.remoteMonitoring.toDateString()})` : t("no"),
+      patientSystem: patient.settings.system ?? trNA,
+      patientRemoteMonitoring,
       timeSpentAwayFromTargetActive,
       frequencyOfSevereHypoglycemiaActive,
       nonDataTransmissionActive,
@@ -270,22 +283,25 @@ function PatientRow(props: PatientElementProps): JSX.Element {
         id={`${rowId}-time-away-target`}
         className={timeSpentAwayFromTargetRateClasses}
       >
-        {`${Math.round(patient.alarm.timeSpentAwayFromTargetRate * 10) /10}%`}
-        {isUserHcp && patientIsMonitored && timeSpentAwayFromTargetActive && <AnnouncementIcon className={classes.alertIcon} />}
+        {`${Math.round(patient.metadata.alarm.timeSpentAwayFromTargetRate * 10) / 10}%`}
+        {isUserHcp && patientIsMonitored && timeSpentAwayFromTargetActive &&
+          <AnnouncementIcon className={classes.alertIcon} />}
       </StyledTableCell>
       <StyledTableCell
         id={`${rowId}-hypo-frequency-rate`}
         className={frequencyOfSevereHypoglycemiaRateClasses}
       >
-        {`${Math.round(patient.alarm.frequencyOfSevereHypoglycemiaRate * 10) /10}%`}
-        {isUserHcp && patientIsMonitored && frequencyOfSevereHypoglycemiaActive && <AnnouncementIcon className={classes.alertIcon} />}
+        {`${Math.round(patient.metadata.alarm.frequencyOfSevereHypoglycemiaRate * 10) / 10}%`}
+        {isUserHcp && patientIsMonitored && frequencyOfSevereHypoglycemiaActive &&
+          <AnnouncementIcon className={classes.alertIcon} />}
       </StyledTableCell>
       <StyledTableCell
         id={`${rowId}-data-not-transferred`}
         className={dataNotTransferredRateClasses}
       >
-        {`${Math.round(patient.alarm.nonDataTransmissionRate * 10) /10}%`}
-        {isUserHcp && patientIsMonitored && nonDataTransmissionActive && <AnnouncementIcon className={classes.alertIcon} />}
+        {`${Math.round(patient.metadata.alarm.nonDataTransmissionRate * 10) / 10}%`}
+        {isUserHcp && patientIsMonitored && nonDataTransmissionActive &&
+          <AnnouncementIcon className={classes.alertIcon} />}
       </StyledTableCell>
       <StyledTableCell id={`${rowId}-ldu`} className={classes.typography}>
         {lastUpload}