+++ packages/yourloops/lib/team/api.ts
@@ -31,12 +31,11 @@ import bows from "bows";
 import { HttpHeaderKeys, HttpHeaderValues } from "../../models/api";
 import { UserRoles } from "../../models/shoreline";
 import { INotificationAPI } from "../../models/notification";
-import { TeamType, ITeam, ITeamMember, TypeTeamMemberRole } from "../../models/team";
+import { ITeam, ITeamMember, TeamType, TypeTeamMemberRole } from "../../models/team";
 import { errorFromHttpStatus } from "../../lib/utils";
 import { Session } from "../auth";
 import appConfig from "../config";
 import { getCurrentLang } from "../language";
-import { PatientMonitored } from "../data/patient";
 import { Monitoring } from "../../models/monitoring";
 
 const log = bows("TeamAPI");
@@ -390,30 +389,6 @@ async function joinTeam(session: Session, teamId: string): Promise<void> {
   return Promise.reject(errorFromHttpStatus(response, log));
 }
 
-async function getMonitoredPatient(session: Session, patientId?: string): Promise<PatientMonitored|null> {
-  const { sessionToken, traceToken, user } = session;
-  log.info("getMonitoredPatient()");
-  const requestedPatientId = patientId? patientId : user.userid;
-
-  const apiURL = new URL(`crew/v0/patients/${requestedPatientId}/monitoring`, appConfig.API_HOST);
-  const response = await fetch(apiURL.toString(), {
-    method: "GET",
-    headers: {
-      [HttpHeaderKeys.traceToken]: traceToken,
-      [HttpHeaderKeys.sessionToken]: sessionToken,
-    },
-  });
-
-  if (response.ok) {
-    return response.json() as Promise<PatientMonitored>;
-  }
-  if (response.status === 404) {
-    return null;
-  }
-
-  return Promise.reject(errorFromHttpStatus(response, log));
-}
-
 export default {
   fetchTeams,
   fetchPatients,
@@ -429,5 +404,4 @@ export default {
   changeMemberRole,
   getTeamFromCode,
   joinTeam,
-  getMonitoredPatient,
 };